Send mouse up events to same window as mouse down was.
Previously if you pressed on a button, then dragged to outside the
window, the button would stay hilighted even after you released the
mouse button.

This commit adds tracking so that the window is stored on initial
mouse down event, and is used for all following mouse events until
the button is released.