Add background filling variant of CGraphicsPort::drawText.
Many times we only want a constant background. In that case the old
code fills the background, reads it back, renders the text and then
writes it back. When used with LCD's (instead of framebuffers) this
causes unnecessary delay and screen flicker.

This commit adds a variant of drawText that takes background color,
so that the background and text can both be rendered at one go.
The old functions still function as before.