stm32_otgfsdev: in epin/epout, make switch to IDLE conditional
Previously the driver would switch to IDLE state in epin/epout if
there was not an active in/out transfer.

This seemed to cause premature switches to IDLE state in atleast
two situations:

1) When usbinterrupt was delayed for some reason, such as debug messages.
   Then there were multiple events (in/out/setup) queued, and they might
   not be processed in the same order as they were received.
2) When waiting for DATA OUT in a SETUP request. Then the state is
   EP0STATE_SETUP_OUT, and stm32_ep0out_receive would handle the data.
   However if stm32_epout gets called first, it will see no OUT request
   and set state to IDLE.

This patch makes the epin/epout switch state only if the previous state
is expected.

I'm not sure if these "unexpected state changes to IDLE" represent errors
elsewhere in code. However they currently do occur, and this patch makes
things work for me.

Another way would be to assert the state before switching to idle. Currently
it would cause multiple assert failures for me.