stm32_spi: Check error flags after send, disable while changing bit count.
If the error flags are not cleared, they may corrupt further transfers.
This is because error conditions stop reception, but reading SR after DR
will reset the error flags. Therefore first byte would get lost but further
bytes would get through.

The most proper solution would be to return the errors to the caller, but
there doesn't seem to be an interface for that.

Another issue: The STM32F4 reference manual notes that the bit count should
not be changed while the SPI peripheral is enabled.