Fix CNxTimer stop for repeated timers.
Previously repeated timers were re-enabled after the timer action event.
Consequently, if the action event handler tried to stop the timer, the
request would be ignored.

This patch changes the order of the action so that the timer is re-enabled
before the callback. There is still no risk of re-entrancy, because everything
executes on the USRWORK work queue.

This patch does change the timing a bit, but it should not be critical because
CNxTimer is mostly used for low-frequency events anyway.