Fix null byte handling in strchr().
As mentioned in the strchr() comment and the ISO C99 standard,
"The terminating null byte is considered to be part of the string."

In practise this means that strchr(str, '\0'); should return a pointer
to the end of the string. The old version in NuttX returned NULL instead.

The correct behaviour is relied upon by e.g. minini library.